move init_db_test_context to test-utils
