+++ kld/Cargo.toml
@@ -73,7 +73,6 @@ serde_derive = "1"
 test-utils = { path = "../test-utils" }
 criterion = { version = "0.5.1", features = ["async_tokio"] }
 bincode = "1.3.3"
-tempfile = "3"
 
 [build-dependencies]
 clap = { version = "4.4", features = ["derive", "env"] }
